pyproject.toml
[project]
name = "claims-automaton"
version = "0.1.0"
description = "AI-powered claims processing system using Semantic Kernel multi-agent orchestration"
readme = "README.md"
authors = [
    { name = "Your Name" }
]
requires-python = ">=3.10"
dependencies = [
    "semantic-kernel>=1.12.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pandas>=2.2.0",
    "openai>=1.58.1",
    "azure-identity>=1.17.1",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "opentelemetry-instrumentation-logging>=0.49b0",
    "opentelemetry-distro>=0.49b0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["platforms/semantic-kernel/src"]

[tool.uv]
package = true
index-url = "https://pypi.org/simple"

[tool.uv.pip]
extra-args = []
